Bricksync gradio/requirements.txt
numpy==1.24.3
Werkzeug==2.2.2
pillow==9.5.0
moviepy==1.0.3
allosaurus==1.0.2
pymongo==4.7.0
python-dotenv==1.0.1